refactor(inlayhints): clean up method-call collectors (sonar)
Boy-scout pass over the merged method-call inlay hints:

- Replace the shared abstract base with a MethodCallInlayHintFlags utility
  (composition over inheritance; removes the abstract-class and scoped-bean
  single-arg-ctor smells).
- Split PlatformMethodCallInlayHintCollector into resolution (collector) and
  rendering (PlatformMethodCallHintRenderer): cuts class dependencies and
  method complexity below thresholds, and splits getInlayHints into focused
  steps (fewer returns, single loop exit).
- Drop redundant non-null checks on ANTLR getText(), the unused
  DEFAULT_SHOW_ALL_PARAMETERS field and stale TODOs, name the identifier-center
  divisor constant, and bound the source-defined loop with Math.min instead of
  a break.
- Parameterize four structurally identical "no hints" tests.

Behavior is unchanged; the full inlay-hint test suite stays green.
