+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/inlayhints/PlatformMethodCallInlayHintCollector.java
@@ -21,30 +21,23 @@
  */
 package com.github._1c_syntax.bsl.languageserver.inlayhints;
 
-import com.github._1c_syntax.bsl.languageserver.configuration.Language;
-import com.github._1c_syntax.bsl.languageserver.configuration.LanguageServerConfiguration;
 import com.github._1c_syntax.bsl.languageserver.context.DocumentContext;
 import com.github._1c_syntax.bsl.languageserver.types.TypeService;
 import com.github._1c_syntax.bsl.languageserver.types.model.MemberDescriptor;
 import com.github._1c_syntax.bsl.languageserver.types.model.MemberKind;
-import com.github._1c_syntax.bsl.languageserver.types.model.ParameterDescriptor;
 import com.github._1c_syntax.bsl.languageserver.types.model.SignatureDescriptor;
 import com.github._1c_syntax.bsl.languageserver.types.model.TypeKind;
 import com.github._1c_syntax.bsl.languageserver.types.model.TypeSet;
 import com.github._1c_syntax.bsl.languageserver.types.util.SignatureSelection;
 import com.github._1c_syntax.bsl.languageserver.utils.Ranges;
-import com.github._1c_syntax.bsl.languageserver.configuration.Resources;
 import com.github._1c_syntax.bsl.languageserver.utils.Trees;
 import com.github._1c_syntax.bsl.parser.BSLParser;
 import org.antlr.v4.runtime.Token;
 import org.antlr.v4.runtime.tree.TerminalNode;
-import org.apache.commons.lang3.Strings;
 import org.eclipse.lsp4j.InlayHint;
-import org.eclipse.lsp4j.InlayHintKind;
 import org.eclipse.lsp4j.InlayHintParams;
-import org.eclipse.lsp4j.MarkupContent;
-import org.eclipse.lsp4j.MarkupKind;
 import org.eclipse.lsp4j.Position;
+import org.eclipse.lsp4j.Range;
 import org.jspecify.annotations.Nullable;
 import org.springframework.stereotype.Component;
 
@@ -57,8 +50,10 @@ import java.util.List;
  * но для не-source-defined символов: {@code СтрНайти("a","b","",1)},
  * {@code Сообщение.Сообщить()}, {@code Новый Массив(5)} и т.п.).
  * <p>
- * Коллектор используется единым {@link MethodCallInlayHintSupplier}. Резолв члена
- * выполняется через {@link TypeService#memberAt}, что покрывает три кейса:
+ * Коллектор используется единым {@link MethodCallInlayHintSupplier} и отвечает за
+ * резолв вызова; построение меток и tooltip делегируется
+ * {@link PlatformMethodCallHintRenderer}. Резолв члена выполняется через
+ * {@link TypeService#memberAt}, что покрывает три кейса:
  * <ul>
  *   <li>{@link BSLParser.GlobalMethodCallContext} — глобальная функция
  *       (СтрНайти, Сообщить, ПолучитьСообщенияПользователю …);</li>
@@ -74,27 +69,22 @@ import java.util.List;
  * MemberDescriptor с непустым sourceSymbol.
  */
 @Component
-public class PlatformMethodCallInlayHintCollector
-  extends AbstractMethodCallInlayHintCollector {
+public class PlatformMethodCallInlayHintCollector {
+
+  /** Делитель длины идентификатора для позиционирования курсора в его середине. */
+  private static final int IDENTIFIER_CENTER_DIVISOR = 2;
 
   private final TypeService typeService;
-  private final Resources resources;
+  private final PlatformMethodCallHintRenderer renderer;
 
   public PlatformMethodCallInlayHintCollector(
-    LanguageServerConfiguration configuration,
     TypeService typeService,
-    Resources resources
+    PlatformMethodCallHintRenderer renderer
   ) {
-    super(configuration);
     this.typeService = typeService;
-    this.resources = resources;
+    this.renderer = renderer;
   }
 
-  private Language currentLanguage() {
-    return configuration.getLanguage();
-  }
-
-  @Override
   public List<InlayHint> getInlayHints(DocumentContext documentContext, InlayHintParams params) {
     var ast = documentContext.getAst();
     if (ast == null) {
@@ -103,45 +93,78 @@ public class PlatformMethodCallInlayHintCollector
     var range = params.getRange();
     var result = new ArrayList<InlayHint>();
     for (var node : Trees.findAllRuleNodes(ast, BSLParser.RULE_doCall)) {
-      if (!(node instanceof BSLParser.DoCallContext doCall)) {
-        continue;
+      if (node instanceof BSLParser.DoCallContext doCall) {
+        collectForDoCall(result, documentContext, range, doCall);
       }
-      var paramList = doCall.callParamList();
-      if (paramList == null || paramList.callParam().isEmpty()) {
-        continue;
-      }
-      // Range-filter по позиции открывающей скобки вызова — дешёво.
-      if (!Ranges.containsPosition(range, positionOf(doCall.getStart()))) {
-        continue;
-      }
-      var args = paramList.callParam();
-      // Конструкторы (Новый Тип(...)) резолвятся отдельно — memberAt
-      // не вернёт MemberDescriptor для имени типа.
-      if (doCall.getParent() instanceof BSLParser.NewExpressionContext nex) {
-        appendConstructorHints(result, documentContext, nex, args);
-        continue;
-      }
-      var methodNamePosition = methodNamePosition(doCall);
-      if (methodNamePosition == null) {
-        continue;
-      }
-      var member = typeService.memberAt(documentContext, methodNamePosition)
-        .map(TypeService.TypedMember::descriptor)
-        .filter(m -> m.kind() == MemberKind.METHOD)
-        // Source-defined методы покрывает SourceDefinedMethodCallInlayHintCollector.
-        .filter(m -> m.sourceSymbol() == null)
-        .orElse(null);
-      if (member == null || member.signatures().isEmpty()) {
-        continue;
-      }
-      var argTypes = inferArgTypes(documentContext, args);
-      var signature = pickSignatureByTypes(member.signatures(), args.size(), argTypes);
-      if (signature == null) {
-        continue;
-      }
-      appendHints(result, member, signature, args);
     }
     return result;
+  }
+
+  private void collectForDoCall(
+    List<InlayHint> sink,
+    DocumentContext documentContext,
+    Range range,
+    BSLParser.DoCallContext doCall
+  ) {
+    var args = hintableArguments(doCall, range);
+    if (args == null) {
+      return;
+    }
+    // Конструкторы (Новый Тип(...)) резолвятся отдельно — memberAt
+    // не вернёт MemberDescriptor для имени типа.
+    if (doCall.getParent() instanceof BSLParser.NewExpressionContext nex) {
+      appendConstructorHints(sink, documentContext, nex, args);
+      return;
+    }
+    appendResolvedMethodHints(sink, documentContext, doCall, args);
+  }
+
+  /**
+   * Фактические аргументы вызова, если он вообще подлежит подсказкам: есть непустой
+   * список аргументов и открывающая скобка попадает в запрошенный диапазон.
+   * {@code null}, если подсказки для вызова не нужны.
+   */
+  @Nullable
+  private static List<? extends BSLParser.CallParamContext> hintableArguments(
+    BSLParser.DoCallContext doCall,
+    Range range
+  ) {
+    var paramList = doCall.callParamList();
+    if (paramList == null || paramList.callParam().isEmpty()) {
+      return null;
+    }
+    // Range-filter по позиции открывающей скобки вызова — дешёво.
+    if (!Ranges.containsPosition(range, positionOf(doCall.getStart()))) {
+      return null;
+    }
+    return paramList.callParam();
+  }
+
+  private void appendResolvedMethodHints(
+    List<InlayHint> sink,
+    DocumentContext documentContext,
+    BSLParser.DoCallContext doCall,
+    List<? extends BSLParser.CallParamContext> args
+  ) {
+    var methodNamePosition = methodNamePosition(doCall);
+    if (methodNamePosition == null) {
+      return;
+    }
+    var member = typeService.memberAt(documentContext, methodNamePosition)
+      .map(TypeService.TypedMember::descriptor)
+      .filter(m -> m.kind() == MemberKind.METHOD)
+      // Source-defined методы покрывает SourceDefinedMethodCallInlayHintCollector.
+      .filter(m -> m.sourceSymbol() == null)
+      .orElse(null);
+    if (member == null || member.signatures().isEmpty()) {
+      return;
+    }
+    var argTypes = inferArgTypes(documentContext, args);
+    var signature = pickSignatureByTypes(member.signatures(), args.size(), argTypes);
+    if (signature == null) {
+      return;
+    }
+    renderer.appendHints(sink, member, signature, args);
   }
 
   /**
@@ -155,8 +178,7 @@ public class PlatformMethodCallInlayHintCollector
   ) {
     var result = new ArrayList<TypeSet>(args.size());
     for (var arg : args) {
-      var text = arg.getText();
-      if (text == null || text.isBlank()) {
+      if (arg.getText().isBlank()) {
         result.add(TypeSet.EMPTY);
         continue;
       }
@@ -184,22 +206,16 @@ public class PlatformMethodCallInlayHintCollector
     BSLParser.NewExpressionContext nex,
     List<? extends BSLParser.CallParamContext> args
   ) {
-    if (nex.typeName() == null) {
-      return;
-    }
-    var typeName = nex.typeName().getText();
+    var typeName = nex.typeName() == null ? null : nex.typeName().getText();
     if (typeName == null || typeName.isBlank()) {
       return;
     }
     var fileType = documentContext.getFileType();
     var ref = typeService.resolve(typeName, fileType).orElse(null);
-    if (ref == null) {
-      return;
-    }
     // Конструкторы source-defined типов (OneScript-классы, ПриСозданииОбъекта)
     // покрывает SourceDefinedMethodCallInlayHintCollector — иначе подсказки
     // имён параметров задвоились бы. Здесь обрабатываем только платформенные.
-    if (ref.kind() == TypeKind.USER) {
+    if (ref == null || ref.kind() == TypeKind.USER) {
       return;
     }
     var constructors = typeService.getConstructors(ref, fileType);
@@ -212,7 +228,7 @@ public class PlatformMethodCallInlayHintCollector
       return;
     }
     var synthetic = MemberDescriptor.method(typeName, "", constructors);
-    appendHints(sink, synthetic, signature, args);
+    renderer.appendHints(sink, synthetic, signature, args);
   }
 
   /**
@@ -250,7 +266,7 @@ public class PlatformMethodCallInlayHintCollector
     // Внутрь токена, чтобы findTerminalNodeContainsPosition попал в IDENTIFIER.
     var col = token.getCharPositionInLine();
     var len = token.getText().length();
-    return new Position(token.getLine() - 1, col + Math.max(0, len / 2));
+    return new Position(token.getLine() - 1, col + Math.max(0, len / IDENTIFIER_CENTER_DIVISOR));
   }
 
   private static Position positionOf(Token token) {
@@ -278,146 +294,6 @@ public class PlatformMethodCallInlayHintCollector
       idx = SignatureSelection.pickIndexByArity(signatures, callArgCount);
     }
     return signatures.get(idx < 0 ? 0 : idx);
-  }
-
-  private void appendHints(
-    List<InlayHint> sink,
-    MemberDescriptor member,
-    SignatureDescriptor signature,
-    List<? extends BSLParser.CallParamContext> args
-  ) {
-    var parameters = signature.parameters();
-    if (parameters.isEmpty()) {
-      return;
-    }
-    var lang = currentLanguage();
-    var variadicIndex = variadicIndex(parameters);
-    // Единственный пустой callParam — это запись вида `Метод()` / `Новый Тип()`,
-    // т.е. ноль фактических аргументов. Пропуск аргумента (`Метод(а,,б)`) даёт
-    // несколько callParam'ов, в которых пустой обозначает именно опущенный довод.
-    var hasSkippedArguments = args.size() > 1;
-    for (var i = 0; i < args.size(); i++) {
-      var unit = paramAt(parameters, variadicIndex, i, lang);
-      if (unit == null) {
-        break;
-      }
-      appendHint(sink, member, unit, args.get(i), hasSkippedArguments);
-    }
-  }
-
-  /** Имя параметра + дескриптор для конкретной позиции аргумента. */
-  private record NamedParam(String name, ParameterDescriptor descriptor) {
-  }
-
-  /**
-   * Параметр для позиции аргумента {@code i}. Вариадик-хвост разворачивается в
-   * нумерованные имена по фактическим аргументам (Значение → Значение1, …).
-   * {@code null}, если аргументов больше, чем параметров, и хвост не вариадик.
-   */
-  @Nullable
-  private static NamedParam paramAt(List<ParameterDescriptor> parameters, int variadicIndex,
-                                    int i, Language lang) {
-    if (variadicIndex >= 0 && i >= variadicIndex) {
-      var p = parameters.get(variadicIndex);
-      return new NamedParam(p.displayName(lang) + (i - variadicIndex + 1), p);
-    }
-    if (i < parameters.size()) {
-      var p = parameters.get(i);
-      return new NamedParam(p.displayName(lang), p);
-    }
-    return null;
-  }
-
-  private void appendHint(List<InlayHint> sink, MemberDescriptor member, NamedParam unit,
-                          BSLParser.CallParamContext callParam, boolean argumentSkipAllowed) {
-    var passedValue = callParam.getText();
-    // Пустой довод — либо `Метод()`/`Новый Тип()` (ноль аргументов, хинт не нужен),
-    // либо пропущенный аргумент `Метод(а,,б)`. Для пропущенного при включённом
-    // showDefaultValues показываем значение по умолчанию из сигнатуры.
-    if (passedValue.isBlank()) {
-      if (!argumentSkipAllowed || !showDefaultValues() || unit.descriptor().defaultValue().isBlank()) {
-        return;
-      }
-    } else if (!showParametersWithTheSameName()
-      && shadowsName(passedValue, unit.name(), unit.descriptor())) {
-      return;
-    } else {
-      // непустой довод, не дублирующий имя параметра — показываем обычный хинт ниже
-    }
-    var hint = new InlayHint();
-    hint.setKind(InlayHintKind.Parameter);
-    hint.setLabel(buildLabel(unit.name(), unit.descriptor(), passedValue));
-    hint.setPosition(positionOf(callParam.getStart()));
-    if (!unit.name().isBlank() || !unit.descriptor().description().isBlank()) {
-      hint.setTooltip(buildTooltip(unit.name(), unit.descriptor(), member));
-    }
-    hint.setPaddingRight(Boolean.TRUE);
-    sink.add(hint);
-  }
-
-  /** Индекс вариадик-параметра (он же последний), либо {@code -1}. */
-  private static int variadicIndex(List<ParameterDescriptor> parameters) {
-    for (var i = 0; i < parameters.size(); i++) {
-      if (parameters.get(i).variadic()) {
-        return i;
-      }
-    }
-    return -1;
-  }
-
-  /** Аргумент уже содержит имя параметра — хинт был бы избыточен. */
-  private static boolean shadowsName(String passedValue, String paramName, ParameterDescriptor parameter) {
-    if (paramName.isBlank()) {
-      return false;
-    }
-    var bn = parameter.bilingualName();
-    return Strings.CI.contains(passedValue, paramName)
-      || (!bn.ru().isEmpty() && Strings.CI.contains(passedValue, bn.ru()))
-      || (!bn.en().isEmpty() && Strings.CI.contains(passedValue, bn.en()));
-  }
-
-  private String buildLabel(String paramName, ParameterDescriptor parameter, String passedValue) {
-    var sb = new StringBuilder();
-    sb.append(paramName);
-    if (showDefaultValues() && passedValue.isBlank() && !parameter.defaultValue().isBlank()) {
-      sb.append(" (").append(parameter.defaultValue()).append(')');
-    } else {
-      sb.append(':');
-    }
-    return sb.toString();
-  }
-
-  private MarkupContent buildTooltip(String paramName, ParameterDescriptor parameter, MemberDescriptor member) {
-    var sb = new StringBuilder();
-    var lang = currentLanguage();
-    sb.append("**").append(paramName).append("**");
-    if (parameter.optional()) {
-      sb.append(" _(").append(tr("optionalParameter")).append(")_");
-    }
-    if (!parameter.defaultValue().isBlank()) {
-      sb.append(" _= ").append(parameter.defaultValue()).append('_');
-    }
-    if (!parameter.types().isEmpty()) {
-      sb.append(": ");
-      var first = true;
-      for (var ref : parameter.types().refs()) {
-        if (!first) {
-          sb.append(" | ");
-        }
-        sb.append(ref.qualifiedName());
-        first = false;
-      }
-    }
-    var pDesc = parameter.displayDescription(lang);
-    if (!pDesc.isBlank()) {
-      sb.append("\n\n").append(pDesc);
-    }
-    sb.append("\n\n_").append(tr("methodLabel")).append("_ `").append(member.displayName(lang)).append('`');
-    return new MarkupContent(MarkupKind.MARKDOWN, sb.toString());
-  }
-
-  private String tr(String key) {
-    return resources.getResourceString(getClass(), key);
   }
 
 }
