+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/inlayhints/SourceDefinedMethodCallInlayHintCollector.java
@@ -61,12 +61,9 @@ import java.util.List;
  * разрешения, у хинта нет, поэтому он не несёт {@code data}.
  */
 @Component
-public class SourceDefinedMethodCallInlayHintCollector
-  extends AbstractMethodCallInlayHintCollector {
+public class SourceDefinedMethodCallInlayHintCollector {
 
-  // TODO: высчитать позицию хинта относительно последнего параметра.
-  private static final boolean DEFAULT_SHOW_ALL_PARAMETERS = false;
-
+  private final LanguageServerConfiguration configuration;
   private final ReferenceIndex referenceIndex;
   private final DescriptionFormatter descriptionFormatter;
 
@@ -75,12 +72,11 @@ public class SourceDefinedMethodCallInlayHintCollector
     ReferenceIndex referenceIndex,
     DescriptionFormatter descriptionFormatter
   ) {
-    super(configuration);
+    this.configuration = configuration;
     this.referenceIndex = referenceIndex;
     this.descriptionFormatter = descriptionFormatter;
   }
 
-  @Override
   public List<InlayHint> getInlayHints(DocumentContext documentContext, InlayHintParams params) {
     var range = params.getRange();
     var references = referenceIndex.getReferencesFrom(documentContext.getUri(), SymbolKind.Method).stream()
@@ -120,19 +116,17 @@ public class SourceDefinedMethodCallInlayHintCollector
     var targetUri = methodSymbol.getOwner().getUri().toString();
 
     var hints = new ArrayList<InlayHint>();
-    for (var i = 0; i < parameters.size(); i++) {
-
-      // todo: show all parameters (in config)?
-      if (callParams.size() < i + 1) {
-        break;
-      }
+    // Подсказки — только для фактически переданных аргументов (по последний из них).
+    var hintableCount = Math.min(parameters.size(), callParams.size());
+    for (var i = 0; i < hintableCount; i++) {
 
       var parameter = parameters.get(i);
       var callParam = callParams.get(i);
 
       var passedValue = callParam.getText();
 
-      if (!showParametersWithTheSameName() && Strings.CI.contains(passedValue, parameter.getName())) {
+      if (!MethodCallInlayHintFlags.showParametersWithTheSameName(configuration)
+        && Strings.CI.contains(passedValue, parameter.getName())) {
         continue;
       }
 
@@ -161,7 +155,7 @@ public class SourceDefinedMethodCallInlayHintCollector
     var labelBuilder = new StringBuilder();
     labelBuilder.append(parameter.getName());
 
-    if (showDefaultValues()
+    if (MethodCallInlayHintFlags.showDefaultValues(configuration)
       && passedValue.isBlank()
       && !defaultValue.equals(ParameterDefinition.DefaultValue.EMPTY)
     ) {
