+++ src/main/resources/com/github/_1c_syntax/bsl/languageserver/inlayhints/PlatformMethodCallHintRenderer_en.properties
